refactored based on comments
