+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/sink/cdc/CdcMultiTableParsingProcessFunction.java
@@ -18,15 +18,9 @@
 
 package org.apache.paimon.flink.sink.cdc;
 
-import org.apache.paimon.catalog.Catalog;
-import org.apache.paimon.catalog.Identifier;
-import org.apache.paimon.flink.action.cdc.mysql.MySqlDatabaseSyncMode;
-import org.apache.paimon.table.FileStoreTable;
 import org.apache.paimon.types.DataField;
 
-import org.apache.flink.api.common.typeinfo.TypeHint;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
-import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.api.java.typeutils.ListTypeInfo;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.streaming.api.functions.ProcessFunction;
@@ -36,8 +30,6 @@ import org.apache.flink.util.OutputTag;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * A {@link ProcessFunction} to parse CDC change event to either a list of {@link DataField}s or
@@ -50,44 +42,13 @@ import java.util.stream.Collectors;
 public class CdcMultiTableParsingProcessFunction<T> extends ProcessFunction<T, Void> {
 
     private final EventParser.Factory<T> parserFactory;
-    private final Set<String> initialTables;
-    private final String database;
-    private final Catalog.Loader catalogLoader;
-    private final MySqlDatabaseSyncMode mode;
 
     private transient EventParser<T> parser;
-    private transient Catalog catalog;
     private transient Map<String, OutputTag<List<DataField>>> updatedDataFieldsOutputTags;
     private transient Map<String, OutputTag<CdcRecord>> recordOutputTags;
-    public static final OutputTag<CdcMultiplexRecord> DYNAMIC_OUTPUT_TAG =
-            new OutputTag<>("paimon-dynamic-table", TypeInformation.of(CdcMultiplexRecord.class));
-    public static final OutputTag<Tuple2<Identifier, List<DataField>>>
-            DYNAMIC_SCHEMA_CHANGE_OUTPUT_TAG =
-                    new OutputTag<>(
-                            "paimon-dynamic-table-schema-change",
-                            TypeInformation.of(
-                                    new TypeHint<Tuple2<Identifier, List<DataField>>>() {}));
 
-    public CdcMultiTableParsingProcessFunction(
-            String database,
-            Catalog.Loader catalogLoader,
-            List<FileStoreTable> tables,
-            EventParser.Factory<T> parserFactory) {
-        this(database, catalogLoader, tables, parserFactory, MySqlDatabaseSyncMode.STATIC);
-    }
-
-    public CdcMultiTableParsingProcessFunction(
-            String database,
-            Catalog.Loader catalogLoader,
-            List<FileStoreTable> tables,
-            EventParser.Factory<T> parserFactory,
-            MySqlDatabaseSyncMode mode) {
-        // for now, only support single database
-        this.database = database;
-        this.catalogLoader = catalogLoader;
-        this.initialTables = tables.stream().map(FileStoreTable::name).collect(Collectors.toSet());
+    public CdcMultiTableParsingProcessFunction(EventParser.Factory<T> parserFactory) {
         this.parserFactory = parserFactory;
-        this.mode = mode;
     }
 
     @Override
@@ -95,58 +56,18 @@ public class CdcMultiTableParsingProcessFunction<T> extends ProcessFunction<T, V
         parser = parserFactory.create();
         updatedDataFieldsOutputTags = new HashMap<>();
         recordOutputTags = new HashMap<>();
-        catalog = catalogLoader.load();
     }
 
     @Override
     public void processElement(T raw, Context context, Collector<Void> collector) throws Exception {
         parser.setRawEvent(raw);
-
-        // CDC Ingestion only supports single database at this time being.
-        //    In the future, there will be a mapping between source databases
-        //    and target paimon databases
-        String databaseName = parser.parseDatabaseName();
         String tableName = parser.parseTableName();
-
-        // check for newly added table
-        parser.parseNewTable(database)
-                .ifPresent(
-                        schema -> {
-                            Identifier identifier =
-                                    new Identifier(database, parser.parseTableName());
-                            try {
-                                catalog.createTable(identifier, schema, true);
-                            } catch (Throwable ignored) {
-                            }
-                        });
-
         List<DataField> schemaChange = parser.parseSchemaChange();
         if (schemaChange.size() > 0) {
-            if (mode == MySqlDatabaseSyncMode.DYNAMIC) {
-                context.output(
-                        DYNAMIC_SCHEMA_CHANGE_OUTPUT_TAG,
-                        Tuple2.of(Identifier.create(database, tableName), schemaChange));
-            } else {
-                context.output(getUpdatedDataFieldsOutputTag(tableName), schemaChange);
-            }
+            context.output(getUpdatedDataFieldsOutputTag(tableName), schemaChange);
         }
-
-        if (mode == MySqlDatabaseSyncMode.DYNAMIC) {
-            parser.parseRecords()
-                    .forEach(
-                            record ->
-                                    context.output(
-                                            DYNAMIC_OUTPUT_TAG,
-                                            wrapRecord(database, tableName, record)));
-
-        } else {
-            parser.parseRecords()
-                    .forEach(record -> context.output(getRecordOutputTag(tableName), record));
-        }
-    }
-
-    private CdcMultiplexRecord wrapRecord(String databaseName, String tableName, CdcRecord record) {
-        return CdcMultiplexRecord.fromCdcRecord(databaseName, tableName, record);
+        parser.parseRecords()
+                .forEach(record -> context.output(getRecordOutputTag(tableName), record));
     }
 
     private OutputTag<List<DataField>> getUpdatedDataFieldsOutputTag(String tableName) {
@@ -162,10 +83,6 @@ public class CdcMultiTableParsingProcessFunction<T> extends ProcessFunction<T, V
     private OutputTag<CdcRecord> getRecordOutputTag(String tableName) {
         return recordOutputTags.computeIfAbsent(
                 tableName, CdcMultiTableParsingProcessFunction::createRecordOutputTag);
-    }
-
-    private boolean isTableNewlyAdded(String tableName) {
-        return !initialTables.contains(tableName);
     }
 
     public static OutputTag<CdcRecord> createRecordOutputTag(String tableName) {
